added get_defense
get_defense emulates the dex_app table, in which this allows us to remove the 'maximum' of a table.  So if something went above the max of 25, it would crash the mud, with this (being the first of many) will allow us to break these 'maximums' and allow for infinate caps should we wish to add one.  Breaking Tabled values for more awesomesauce functors like this will save the day!