Update Makefile
Updated the way the mud builds Version.h every-build.  This equates to more details and easily accessable data based on the compiler build.  Including the last successful build with backup name, the engine version, the engine name(InfectEngine), as well as the new directory for commands (Code/commands/*.cpp) which is being developed so we can move each and every cmd_function into their own .cpp file for ease of debugging the files.